+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.0.2
+* Fix: selling relics no longer buys the card behind the click
+* Fix: work around basemod hack that made selling potions seem impossible
+
 ### v1.0.1
 * Fix relic description